implement rules for juxtaposition
there need to be some rules how certain groups of syntactic expressions
can be juxtaposed so there are no ambiguities
e.g. it should not be possible to have two consecutive
ints (+123-456) or identifiers (..abc.def)
so types are split inot two syntactic groups: enclosed and non-enclosed
enclosed ones are the one with a start and end delimiter
e.g. text, form, tuple, record
the rest is non-enclosed
the rules for juxtaposition of expressions are:
- after an enclosed expressions anything can follow
- after a non-enclosed expressions there must be a whitespace or
  an enclosed expression